requirements_32bit.txt
python-telegram-bot
python-dotenv
fake_useragent
pyyaml
pymongo
python-dotenv
pandas
requests
appdirs
tabulate
lxml
tabulate
beautifulsoup4
html5lib
PyQt5
psutil
pywin32
sqlalchemy
mariadb
pymysql
stem
openpyxl